[UTIL] fixing unit test

+++ test/test_util.py
@@ -80,7 +80,7 @@ def test_restriction():
     state_sequence = np.array([[0, i] for i in range(M)] * 10, dtype=np.intc)
     bias_energy_sequence = np.array([[i] * T for i in range(X)], dtype=np.float64)
     cset = [i for i in range(M) if i % 2 == 0]
-    ref_state_sequence = np.array([[0, i] for i in range(M / 2)] * 10, dtype=np.intc)
+    ref_state_sequence = np.array([[0, i] for i in range(int(M / 2))] * 10, dtype=np.intc)
     ref_bias_energy_sequence = np.array([[i] * T for i in range(X) if i % 2 == 0], dtype=np.float64)
     new_state_sequence, new_bias_energy_sequence = util.restrict_samples_to_cset(
         state_sequence, bias_energy_sequence, cset)
